refactoring - some files were moved to separated directories
