+++ src/view/comments/comments-view.js
@@ -1,4 +1,4 @@
-import Smart from "./smart";
+import Smart from "../smart";
 
 const getCommentsTemplate = () => {
   return `<ul class="film-details__comments-list">
