+++ src/view/films/films-board-view.js
@@ -1,4 +1,4 @@
-import AbstractView from "./abstract";
+import AbstractView from "../abstract";
 
 const createFilmsTemplate = () => {
   return `<section class="films">
