+++ src/view/films/films-list-view.js
@@ -1,4 +1,4 @@
-import AbstractView from "./abstract";
+import AbstractView from "../abstract";
 
 const createFilmsListTemplate = () => {
   return `<section class="films-list">
